# contracts/python/models/user.py
# Auto-generated by /data-contract-generator
# @spec FEAT-001/DM-001, FEAT-002/DM-001
# Do not edit manually - regenerate from specs

from datetime import datetime
from uuid import UUID

from pydantic import BaseModel, EmailStr, Field


class UserBase(BaseModel):
    """Base user model with shared attributes."""

    email: EmailStr = Field(..., max_length=255)


class UserCreate(UserBase):
    """Model for user creation request. @spec FEAT-001/AC-001"""

    password: str = Field(..., min_length=12)
    password_confirm: str


class UserInDB(UserBase):
    """Full user model as stored in database. @spec FEAT-001/DM-001"""

    id: UUID
    password_hash: str
    email_verified: bool = False
    created_at: datetime
    updated_at: datetime

    class Config:
        from_attributes = True


class UserResponse(BaseModel):
    """User data returned in API responses. @spec FEAT-002/API-001"""

    id: UUID
    email: EmailStr


class SessionBase(BaseModel):
    """Base session model. @spec FEAT-002/DM-001"""

    user_id: UUID
    user_agent: str | None = None
    ip_address: str | None = None


class SessionInDB(SessionBase):
    """Full session model. @spec FEAT-002/DM-001"""

    id: UUID
    refresh_token_hash: str
    created_at: datetime
    expires_at: datetime
    revoked_at: datetime | None = None

    class Config:
        from_attributes = True


class TokenResponse(BaseModel):
    """Token response for login. @spec FEAT-002/API-001"""

    access_token: str
    refresh_token: str
    token_type: str = "Bearer"
    expires_in: int = 3600
    user: UserResponse


class RefreshTokenRequest(BaseModel):
    """Refresh token request. @spec FEAT-002/API-002"""

    refresh_token: str


class AccessTokenResponse(BaseModel):
    """Access token response. @spec FEAT-002/API-002"""

    access_token: str
    token_type: str = "Bearer"
    expires_in: int = 3600