# specs/api/openapi.yaml
# OpenAPI 3.0 Specification
# Generated from feature specs - do not edit manually
# Run: /interface-first-designer to regenerate

openapi: 3.0.3
info:
  title: SDD Template API
  description: API specification for SDD Template application
  version: 0.1.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.example.com
    description: Production

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Users
    description: User management endpoints

paths:
  # FEAT-001: User Registration
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      description: |
        Create a new user account.

        @spec FEAT-001/AC-001
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address
      description: |
        Verify user's email with token from verification email.

        @spec FEAT-001/AC-004
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # FEAT-002: User Login
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      description: |
        Authenticate user with email and password.

        @spec FEAT-002/AC-001
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'

  /auth/logout:
    post:
      tags: [Auth]
      summary: User logout
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # Request schemas
    RegisterRequest:
      type: object
      required: [email, password, password_confirm]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 12
          example: SecureP@ss123
        password_confirm:
          type: string
          format: password
          example: SecureP@ss123

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Verification token from email

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # Response schemas
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: Registration successful. Please check your email.
        user_id:
          type: string
          format: uuid

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
          description: Seconds until expiry
        user:
          $ref: '#/components/schemas/UserResponse'

    AccessTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string